+++ CHANGELOG.md
@@ -21,3 +21,6 @@
 * (ERICK) Created `isInteger()` method.
 * (ERICK) Created `updateRoomDetails()` method.
 * `cancelReservation()` takes string instead of int.
+
+##### v0.4 [14.03.2021]: 
+* (ERICK) Added javadoc documentation to RoomHandler.